Fixes #10 A.C.T. Not playing nice with SMP.
A.C.T. Causing SMP crashes on placement and usage.

+++ src/main/java/thaumicenergistics/network/packet/client/PacketClientAspectSlot.java
@@ -16,7 +16,7 @@ import thaumicenergistics.network.packet.AbstractPacket;
 public class PacketClientAspectSlot
 	extends AbstractClientPacket
 {
-	private static final int MODE_LIST_UPDATE = 0;
+	private static final byte MODE_LIST_UPDATE = 0;
 
 	private List<Aspect> filterAspects;
