+++ src/main/java/thaumicenergistics/network/packet/server/PacketServerEssentiaTerminal.java
@@ -12,9 +12,9 @@ import thaumicenergistics.parts.AEPartEssentiaTerminal;
 public class PacketServerEssentiaTerminal
 	extends AbstractServerPacket
 {
-	private static final int MODE_SELECTED_ASPECT = 0;
-	private static final int MODE_FULL_UPDATE = 1;
-	private static final int MODE_SORT_CHANGE = 2;
+	private static final byte MODE_SELECTED_ASPECT = 0;
+	private static final byte MODE_FULL_UPDATE = 1;
+	private static final byte MODE_SORT_CHANGE = 2;
 	
 	private static final ComparatorMode[] SORT_MODES = ComparatorMode.values();
